Staged login scene to go straight to guardian mode menu regardless of input
The login button in the login scene is currently set up to go straight to the next desired scene.  When the code for user authentication is to be done, we can just have the object sent a message to itself to invoke the scene switch method and eliminate the problem of scene switching.